[LuxTestUtils] Move mooncake to an extension
