+++ lib/LuxTestUtils/Project.toml
@@ -16,14 +16,19 @@ ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 Functors = "d9f16b24-f501-4c13-a1f2-28368ffc5196"
 JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 MLDataDevices = "7e8f7934-dd98-4c1a-8fe8-92b47a384d40"
-Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
 Optimisers = "3bd65402-5787-11e9-1adc-39752487f4e2"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
+[weakdeps]
+Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
+
 [sources]
 MLDataDevices = {path = "../MLDataDevices"}
+
+[extensions]
+MooncakeExt = "Mooncake"
 
 [compat]
 ADTypes = "1.10"
