+++ lib/LuxTestUtils/src/autodiff.jl
@@ -74,26 +74,6 @@ function gradient(::F, ::AutoEnzyme{<:Enzyme.ForwardMode}, args...) where {F}
     return error("AutoEnzyme{ForwardMode} is not supported yet.")
 end
 
-function gradient(f::F, ::AutoMooncake, args...) where {F}
-    return gradient(f, mooncake_gradient_function, args...)
-end
-
-"""
-    mooncake_gradient_function(f, x)
-
-Compute gradient using Mooncake.jl's value_and_gradient!! function.
-Returns only the gradient for args x.
-"""
-function mooncake_gradient_function(f, x)
-    # Enable friendly_tangents for testing.
-    cache = Mooncake.prepare_gradient_cache(
-        f, x; config=Mooncake.Config(; friendly_tangents=true)
-    )
-    y, tangents = Mooncake.value_and_gradient!!(cache, f, x)
-    tangent_func, tangent_args = tangents
-    return tangent_args
-end
-
 # ForwardDiff.jl
 function gradient(f::F, ::AutoForwardDiff, args...) where {F}
     return gradient(f, ForwardDiff.gradient, args...)
